Add files via upload
This python script is intended to be a general purpose model for identifying relevant variables and selecting the best model for prediction/classification. It is designed to work over a broad range of structured data with minimal effort from the user. It is perpetual work in progress and should be treated as such.