Update Model Builder.py
This is meant to be a general purpose machine learning model builder that works with most structured datasets. You just plug in a handful of inputs at the top (lines 20-40) and the model will pair down your data to the most useful inputs, train a handful of different models on the data, and tell you which one is the most accurate along with some other relevant info. Enjoy!