Update Model Builder.py
This is a (mostly complete) model that I have built to automate most of the data preprocessing, dimension reduction, and model training/testing/selection that goes into the initial steps of a machine learning project. My goal is to create a general purpose model that can be quickly applied to most datasets to help data scientists quickly identify the most relevant factors and the most effective models to use for prediction/classification problems.